backend/requirements.txt
fastapi
uvicorn
langchain
langchain-openai
langchain-community
chromadb
python-dotenv
sqlmodel
langchain-chroma
langgraph
slowapi
python-jose[cryptography]
